refactor(sandbox): split vite.ts + nitro/module.ts
